% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-vec.R
\name{norm_vec}
\alias{norm_vec}
\title{Normalise vector with (x - min) / (max- min).}
\usage{
norm_vec(df, col)
}
\arguments{
\item{df}{Dataframe with column to normalise}

\item{col}{Character giving the name of the column to normalise}
}
\value{
Dataframe with the normalised column
}
\description{
Normalise vector with (x - min) / (max- min).
}
\examples{
ssb_ices <- read.csv(system.file("extdata", "ssa-ices-combinded.csv", package = "advancedr"))
ssb_ices$ssb_norm <- norm_vec(ssb_ices, "ssb")
}

