% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-data.R
\name{split_data}
\alias{split_data}
\title{Split dataframe in list of dataframes according to specific column.}
\usage{
split_data(df, col)
}
\arguments{
\item{df}{Dataframe to split}

\item{col}{Character of the column to split by}
}
\value{
List of dataframes
}
\description{
Split dataframe in list of dataframes according to specific column.
}
\examples{
ssb_ices <- read.csv(system.file("extdata", "ssa-ices-combinded.csv", package = "advancedr"))
split_data(ssb_ices, col = "species")
}

